requirements.txt
pytest==9.0.2
selenium==4.41.0
allure-pytest==2.15.3
faker==40.11.0
webdriver-manager==4.0.2
